updated deprecated type_aliases import; renamed to match convention
